==> R/02_clean.qmd <==
---
title: "Clean: Group 03"
format:
  html:
    embed-resources: true
editor: visual
---

## Load libraries

```{r, message=FALSE}
suppressWarnings({
library("tidyverse")
library("here")
library("broom")
})
```

## Clean the meta data

```{r}
suppressWarnings({
#Data is loaded without defined column names
meta_data <- read_csv(here("data/_raw/IDs_mapping.csv"),
                      na = "", 
                      col_names = c("type_id", "description"), 
                      show_col_types = FALSE)


# Remove NA columns and add a column based on admission type, that is numeric. The NA values in this, will be the names of the type of metadata the id describes. Also adds a meta_type column, which has the type in the cells that are NAs in the numeric type_id_as_num
meta_data <- meta_data |> 
  na.omit() |> 
  mutate(type_id, 
         type_id_as_num = as.numeric(type_id)) |> 
  mutate(meta_type = ifelse(is.na(type_id_as_num), 
                            type_id, 
                            NA))

#Fills meta_type with the last non-NA cell 
meta_data <- meta_data |> 
    fill(meta_type)

#Removes the rows containing the names, drops admission_type_is_as_num
meta_data_clean <- meta_data |> 
  filter(!is.na(type_id_as_num)) |> 
  select(!type_id) |> 
  rename(type_id = type_id_as_num)

})

admission_type <- meta_data_clean |> 
  filter(meta_type == "admission_type_id") |> 
  select(-meta_type) |> 
  mutate(description = case_when(description %in% c("NULL", "Not Mapped", "Not Available") ~ NA, 
                                 TRUE ~ description))

discharge_disposition <- meta_data_clean |> 
  filter(meta_type == "discharge_disposition_id") |> 
  select(-meta_type) |> 
  mutate(description = case_when(description %in% c("NULL", "Not Mapped", "Unknown/Invalid") ~ NA, 
                                 TRUE ~ description))

admission_source <- meta_data_clean |> 
  filter(meta_type == "admission_source_id") |> 
  select(-meta_type) |> 
  mutate(description = case_when(description %in% c("NULL", "Not Mapped", "Unknown/Invalid", "Not Available") ~ NA, 
                                 TRUE ~ description))
```

## Clean ICD9 data

The data is downloaded here as the newest version from 2014:

[ICD-9-CM Diagnosis and Procedure Codes: Abbreviated and Full Code Titles \| CMS](https://www.cms.gov/medicare/coding-billing/icd-10-codes/icd-9-cm-diagnosis-procedure-codes-abbreviated-and-full-code-titles)

From the files the short and long ds are chosen (sg are surgical and dx are diagnoses).

When running the 01_load.qmd, all the relevant files should be downloaded and placed in data/\_raw

```{r}
# First the data will be loaded into a dataframe of one column
ICD9_short <- read_delim(here("data/_raw/CMS32_DESC_SHORT_DX.txt"), 
                 col_names = "data", 
                 show_col_types = FALSE, 
                 delim = "\n")

ICD9_long <- read_delim(here("data/_raw/CMS32_DESC_LONG_DX.txt"), 
                 col_names = "data", 
                 show_col_types = FALSE, 
                 delim = "\n")

#The long dataset includes é and è in for example Ménière's disease. To deal with this, the data is converted to UTF-8
ICD9_long <- ICD9_long |> 
  mutate(data = iconv(data, 
        from = "latin1", 
        to = "UTF-8", 
        sub = "byte"))

# The first value in each row is the ID, but the delimiter is not preserved. Each row is therefore split at the first, and everything else is merged into the second column:

ICD9_short <- ICD9_short |> 
  separate(col = data,
           into = c("ID", "Description_short"), 
           extra = "merge")

ICD9_long <- ICD9_long |> 
  separate(col = data,
           into = c("ID", "Description_long"), 
           extra = "merge")

# The long and short dataframe, should have identical IDs, however in case there are some differences, we want all data, and therefore use full_join
ICD9 <- full_join(ICD9_short, 
                 ICD9_long, 
                 by = "ID")

suppressWarnings({
ICD9 <- ICD9 |> 
  mutate(ID = ifelse(str_starts(ID, "[VE]"),
                     ID,
                     as.double(str_c(substr(ID,1,3),
                           ".",
                           substr(ID,4,length(ID))))))
})
```

## Clean diabetes data

```{r}
diabetes_data <- read_csv(here("data/_raw/diabetic_data.csv"), 
                          na = c("?", "None"), 
                          show_col_types = FALSE)

#Change NO in readmitted column to No
diabetes_data <- diabetes_data |> 
  mutate(readmitted, 
         readmitted = str_replace(string = readmitted, 
                                  "NO", 
                                  "No"))

#Change Ch in change column to Yes
diabetes_data <- diabetes_data |> 
  mutate(change, 
         change = str_replace(string = change, 
                                  "Ch", 
                                  "Yes"))

```

## Join information from metadata to diabetes data

```{r}

full_data <- diabetes_data |> 
  left_join(admission_type,
            join_by('admission_type_id' == 'type_id'), 
            relationship = "many-to-one") |>
  rename(admission_type = description) |> 
  relocate(admission_type, .after = admission_type_id) |> 
  select(-admission_type_id)

full_data <- full_data |> 
  left_join(admission_source,
            join_by('admission_source_id' == 'type_id'), 
            relationship = "many-to-one") |>
  rename(admission_source = description) |> 
  relocate(admission_source, .after = admission_source_id) |> 
  select(-admission_source_id)

full_data <- full_data |> 
  left_join(discharge_disposition,
            join_by('discharge_disposition_id' == 'type_id'), 
            relationship = "many-to-one") |>
  rename(discharge_disposition = description) |> 
  relocate(discharge_disposition, .after = discharge_disposition_id) |> 
  select(-discharge_disposition_id)

```

## Join ICD9 to clean data

```{r}
full_data <- full_data |> 
  left_join(ICD9, 
            join_by('diag_1' == 'ID'), 
            relationship = "many-to-one", 
            na_matches = "never") |> 
  rename(diag_1_short = Description_short, 
          diag_1_long = Description_long) |> 
  mutate(diag_1_short = case_when(is.na(diag_1_short) ~ diag_1,
                                  !is.na(diag_1_short) ~ diag_1_short)) |> 
  mutate(diag_1_long = case_when(is.na(diag_1_long) ~ diag_1,
                                 !is.na(diag_1_long) ~ diag_1_long))

full_data <- full_data |> 
  left_join(ICD9, 
            join_by('diag_2' == 'ID'), 
            relationship = "many-to-one", 
            na_matches = "never") |> 
  rename(diag_2_short = Description_short, 
         diag_2_long = Description_long)|> 
  mutate(diag_2_short = case_when(is.na(diag_2_short) ~ diag_2,
                                  !is.na(diag_2_short) ~ diag_2_short)) |> 
  mutate(diag_2_long = case_when(is.na(diag_2_long) ~ diag_2,
                                  !is.na(diag_2_long) ~ diag_2_long))

full_data <- full_data |> 
  left_join(ICD9, 
            join_by('diag_3' == 'ID'), 
            relationship = "many-to-one", 
            na_matches = "never") |> 
  rename(diag_3_short = Description_short,
         diag_3_long = Description_long)|> 
  mutate(diag_3_short = case_when(is.na(diag_3_short) ~ diag_3,
                                  !is.na(diag_3_short) ~ diag_3_short)) |> 
  mutate(diag_3_long = case_when(is.na(diag_3_long) ~ diag_3,
                                  !is.na(diag_3_long) ~ diag_3_long))


full_data <- full_data |> 
  relocate(c(diag_1_short, 
             diag_2_short, 
             diag_3_short, 
             diag_1_long, 
             diag_2_long, 
             diag_3_long), .after = diag_1) |> 
  select(-diag_1, -diag_2, -diag_3)
```

## Output clean data

```{r}
write_csv(full_data, here("data/02_dat_clean.csv"))
```
