==> R/05_visualizations.qmd <==
---
title: "Visulizations: Group 03"
format:
  html:
    embed-resources: true
editor: visual
---

## Libraries

```{r,message=FALSE,warning=FALSE}

library(tidyverse)
library(here)
```

## Load data

```{r}
data <- read.csv(here("data/03_dat_aug.csv"))
```

## Age distribution

```{r average_number_of_diagnoses_for_all}
data |>
  group_by(age, gender) |>
  summarise(average_diagnoses = mean(number_diagnoses), .groups = "drop") |>
ggplot(aes(x = age,
                y = average_diagnoses,
                fill = gender)) +
  geom_col(position = "dodge") +
  labs(x = "Age Group",
       y = "Average number of diagnoses",
       fill = "Gender",
       title = "Average number of diagnoses for all age groups and genders per encounter") +
  scale_fill_manual(values=c('hotpink',
                             'cornflowerblue',
                             'grey35')) 
```

## Age distribution, shown by race

```{r average_diagnoses_for_all_races}
data |>
  mutate(race = ifelse(is.na(race),'Other',race)) |>
  mutate(race = ifelse(race == 'Other', 'Other/Unspecified', race)) |>
  group_by(age, 
           gender, 
           race) |> 
  summarize(average_diagnoses = mean(number_diagnoses), .groups = "drop") |> 
  ggplot(mapping = aes(x = age,
                y = average_diagnoses,
                fill = gender)) +
  geom_col(position = "dodge", 
           alpha = 1.0) +
  facet_wrap(~ race) +
  labs(
    x = "Age Group",
    y = "Average number of diagnoses",
    fill = "Gender",
    title = 
      "Average number of diagnoses for all age groups and genders,\nfacet wrapped for race") +
    scale_fill_manual(values=c('hotpink',
                             'cornflowerblue',
                             'grey35')) +
    theme_minimal() +
    theme(axis.text.x = element_text(angle = 60, hjust =1),
          legend.position = 'inside',
          legend.position.inside = c(0.85,0.15))

```

## Race/gender distribution

```{r race_gender_distribution}
data |>
  select(patient_nbr, race, gender) |>
  distinct(patient_nbr, .keep_all = TRUE) |>
  mutate(race = ifelse(is.na(race), "Other", race)) |>
  group_by(race, gender) |>
  summarise(count = n(), .groups = "drop") |>
  mutate(text_color = ifelse(count > 4000, "white", 
                       ifelse(gender == "Female", "hotpink", 
                        ifelse(gender == "Male", "cornflowerblue", "grey35")))) |>
  ggplot(aes(x = race, y = count, fill = gender)) +
  geom_bar(stat = "identity", position = "stack") +
  geom_text(
    aes(label = count, color = text_color),
    data = data |> 
      select(patient_nbr, race, gender) |>
      distinct(patient_nbr, .keep_all = TRUE) |>
      mutate(race = ifelse(is.na(race), "Other", race)) |>
      group_by(race, gender) |>
      summarise(count = n(), .groups = "drop") |>
      filter(count > 4000) |>
      mutate(text_color = ifelse(count > 4000, "white", 
                           ifelse(gender == "Female", "hotpink", 
                            ifelse(gender == "Male", "cornflowerblue", "grey35")))),
    position = position_stack(vjust = 0.5),
    size = 4
  ) +
  geom_text(
    aes(label = count, color = text_color),
    data = data |> 
      select(patient_nbr, race, gender) |>
      distinct(patient_nbr, .keep_all = TRUE) |>
      mutate(race = ifelse(is.na(race), "Other", race)) |>
      group_by(race, gender) |>
      summarise(count = n(), .groups = "drop") |>
      filter(count <= 4000) |>
      mutate(text_color = ifelse(count > 4000, "white", 
                           ifelse(gender == "Female", "hotpink",
                            ifelse(gender == "Male", "cornflowerblue", "grey35")))),
    position = position_dodge2(width = 0.8),
    size = 4,
    vjust = -3
  ) +
  scale_color_identity() +
  labs(
    x = "Race", 
    y = "Count", 
    fill = "Gender",
    title = str_c("Gender/race distribution of dataset, total no. of patients: ", 
                  data |> 
                    select(patient_nbr) |>
                    group_by(patient_nbr) |>
                    summarise(count = n()) |>
                    count()
                  )
    ) +
  scale_fill_manual(values = c("hotpink", "cornflowerblue", "grey35")) +
  theme_minimal() +
  theme(legend.position = "bottom")
```

## Discharge disposition

```{r discharge_disposition_for_all_admission_types, fig.width=8, fig.height=6}
#Finding the 8 most frequent discharge disposition for any admission type
top_8_discharge_dispositions <- data |>
  group_by(discharge_disposition)|> 
  summarize(n = n()) |> #Summarizing the total amount of occurences of each discharge type
  arrange(desc(n)) |> #Arranging from high -> low 
  head(8) |> 
  select(discharge_disposition) |>
  pull() #Converting into a vector

data |> 
  select(discharge_disposition, age) |> 
  filter(discharge_disposition %in% top_8_discharge_dispositions) |> #Filtering for only the cases where the discharge type was in the top 8 
  group_by(discharge_disposition, age) |> 
  summarize(n = n(),.groups = 'keep') |> 
  drop_na(discharge_disposition) |>
  ggplot(
    mapping = aes(x = discharge_disposition,
                  y = n,
                  fill = age)
  ) +
  geom_col(position = "dodge") +
  scale_fill_viridis_d(option = "viridis") +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 35)) +
  labs(
    x = "Discharge disposition",
    y = "Count",
    fill = "Age group",
    title = "Discharge disposition for all age groups for all admission types"
  ) +
    theme_minimal() +
    theme(
      axis.text.x = element_text(angle = 60, hjust =1)
    )
```

### Only for emergency admission type

```{r discharge_disposition_for_emergency_admissions, fig.width=8, fig.height=6}
#Finding the top 8 for the casses related to emergency admissions only
top_8_discharge_dispositions <- data |>
  filter(admission_type == "Emergency") |> 
  group_by(discharge_disposition)|> 
  summarize(n = n()) |> 
  arrange(desc(n)) |> 
  head(8) |> 
  select(discharge_disposition) |> 
  pull()

data |>
  filter(admission_type == "Emergency") |> 
  select(discharge_disposition, age) |> 
  filter(discharge_disposition %in% top_8_discharge_dispositions) |> 
  group_by(discharge_disposition, age) |> 
  summarize(n = n(),.groups = 'keep') |> 
  drop_na(discharge_disposition) |>
  ggplot(
    mapping = aes(x = discharge_disposition,
                  y = n,
                  fill = age)
  ) +
  geom_col(position = "dodge") +
  scale_fill_viridis_d(option = "viridis") +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 35)) +
  labs(
    x = "Discharge disposition",
    y = "Count",
    fill = "Age group",
    title = "Discharge disposition for all age groups for emergency admissions"
  ) +
    theme_minimal() +
    theme(
      axis.text.x = element_text(angle = 60, hjust =1)
    )
```

### Only urgent admission type:

```{r discharge_disposition_for_urgent_admissions, fig.width=8, fig.height=6}
top_8_discharge_dispositions <- data |>
  filter(admission_type == "Urgent") |> 
  group_by(discharge_disposition)|> 
  summarize(n = n()) |> 
  arrange(desc(n)) |> 
  head(8) |> 
  select(discharge_disposition) |> 
  pull()

data |>
  filter(admission_type == "Urgent") |> 
  select(discharge_disposition, age) |> 
  filter(discharge_disposition %in% top_8_discharge_dispositions) |> 
  group_by(discharge_disposition, age) |> 
  summarize(n = n(),.groups = 'keep') |> 
  drop_na(discharge_disposition) |>
  ggplot(
    mapping = aes(x = discharge_disposition,
                  y = n,
                  fill = age)
  ) +
  geom_col(position = "dodge") +
  scale_fill_viridis_d(option = "viridis") +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 35)) +
  labs(
    x = "Discharge disposition",
    y = "Count",
    fill = "Age group",
    title = "Discharge disposition for all age groups for urgent admissions"
  ) +
    theme_minimal() +
    theme(
      axis.text.x = element_text(angle = 60, hjust =1)
    )
```

### For elective admission type

```{r discharge_disposition_for_elective_admissions, fig.width=8, fig.height=6}
top_8_discharge_dispositions <- data |>
  filter(admission_type == "Elective") |> 
  group_by(discharge_disposition)|> 
  summarize(n = n()) |> 
  arrange(desc(n)) |> 
  head(8) |> 
  select(discharge_disposition) |> 
  pull()

data |>
  filter(admission_type == "Elective") |> 
  select(discharge_disposition, age) |> 
  filter(discharge_disposition %in% top_8_discharge_dispositions) |> 
  group_by(discharge_disposition, age) |> 
  summarize(n = n(),.groups = 'keep') |> 
  drop_na(discharge_disposition) |>
  ggplot(
    mapping = aes(x = discharge_disposition,
                  y = n,
                  fill = age)
  ) +
  geom_col(position = "dodge") +

  scale_fill_viridis_d(option = "viridis") +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 35)) +
  labs(
    x = "Discharge disposition",
    y = "Count",
    fill = "Age group",
    title = "Discharge disposition for all age groups for elective admissions"
  ) +
    theme_minimal() +
    theme(
      axis.text.x = element_text(angle = 60, hjust =1)
    )
```

## Number of visits per patient

```{r number_of_visits_per_patient}
data |>
  select(patient_nbr,gender) |>
  group_by(patient_nbr) |>
  mutate(number_of_visits = n()) |>
  distinct(patient_nbr,.keep_all = TRUE) |>
  ungroup() |>
  select(gender, number_of_visits) |>
  group_by(number_of_visits,gender) |>
  summarise(count = n(), .groups = 'drop') |>
  ggplot(aes(x = number_of_visits, y = gender)) + 
  geom_count(aes(color = gender, 
                 size = count,  
                 shape = factor(count)),
      position = position_jitter(height = 0.2)) +
  labs(x = "Number of visits",
       y = "Gender",
       title = "Number of visits by gender") +
  scale_color_manual(values = c("hotpink", "cornflowerblue", "grey35")) +
  scale_size_continuous(
    range = c(2, 10), 
    trans = 'sqrt',  
    breaks = c(1, 5, 100, 1000, 10000, 20000),
    labels = c('1', '5', '100', '1000', '10000', '20000+')) +
  scale_shape_manual(values=c(17,rep(16, 28))) +  
  theme_minimal() +
  theme(legend.position = 'bottom',
        legend.justification = "center") +
        guides( color = guide_none(),
          size = guide_legend(title = "Counts", 
                        override.aes = list(shape = c(17,rep(16,5))),
                        ncol = 6),  
          shape = guide_none()) +
  coord_flip()
```

## Diagnosis vs. visits

```{r diagnosis_for_1_visit}
data |>
  filter(nr_encounters == 1) |>
  select(patient_nbr, number_diagnoses, gender) |>
  group_by(patient_nbr) |>
  mutate(visit_num = row_number()) |>
  ungroup() |>
  pivot_wider(
    names_from = visit_num,                
    values_from = number_diagnoses,        
    names_prefix = "visit_",          
    values_fill = list(number_diagnoses = NA)) |>
  ggplot(aes(x=visit_1,fill=gender)) +
  geom_boxplot() +
  labs(title = 'Amount of diagnoses for people with only 1 visit',
       x ='Number of diagnoses',
       fill = 'Gender') + 
  scale_fill_manual(values=c('hotpink',
                             'cornflowerblue',
                             'grey35')) 
  
```

```{r diagnosis_for_2_visits}
data |>
  filter(nr_encounters == 2) |>
  select(patient_nbr, number_diagnoses,encounter_id) |>
  group_by(patient_nbr) |>
  arrange(encounter_id, .by_group = TRUE) |>
  mutate(visit_num = row_number()) |>
  ungroup() |>
  select(-encounter_id) |>
  pivot_wider(
    names_from = visit_num,                
    values_from = number_diagnoses,        
    names_prefix = "visit_",          
    values_fill = list(number_diagnoses = NA)) |>
  pivot_longer(cols = -patient_nbr,
               names_to = 'visit') |>
  ggplot(aes(x=value,fill=visit)) +
  geom_boxplot() +
  labs(title = str_c('Amount of diagnoses for people with 2 visits (',
                      data |>
                       filter(nr_encounters == 2) |>
                       pull(patient_nbr) |>
                       unique() |>
                       length(),
                     ' people)'),
       x ='Number of diagnoses',
       fill = 'Visit number') +
  scale_fill_manual(
    values = c("visit_1" = "lightblue3", "visit_2" = "seagreen3"), 
    labels = c("1", "2") 
  )
```

```{r diagnosis_for_3_visits}
data |>
  filter(nr_encounters == 3) |>
  select(patient_nbr, number_diagnoses, encounter_id) |>
  group_by(patient_nbr) |>
  arrange(encounter_id, .by_group = TRUE) |>
  mutate(visit_num = row_number()) |>
  ungroup() |>
  select(-encounter_id) |>
  pivot_wider(
    names_from = visit_num,                
    values_from = number_diagnoses,        
    names_prefix = "visit_",          
    values_fill = list(number_diagnoses = NA)) |>
  pivot_longer(cols = -patient_nbr,
               names_to = 'visit') |>
  ggplot(aes(x=value,fill=visit)) +
  geom_boxplot() +
  labs(title = str_c('Amount of diagnoses for people with 3 visits (',
                     data |>
                       filter(nr_encounters == 3) |>
                       pull(patient_nbr) |>
                       unique() |>
                       length(),
                     ' people)'),
       x ='Number of diagnoses',
       fill = 'Visit number') +
scale_fill_manual(
    values = c("visit_1" = "lightblue3", "visit_2" = "seagreen3","visit_3" = 'darkolivegreen'), 
    labels = c("1", "2","3") 
  )
```

## Plot of average of number of diagnosis per patient per maximum visit:

```{r average_number_of_diagnoses_per_visit_by_patient}
data |>
  select(patient_nbr, number_diagnoses,encounter_id) |>
  group_by(patient_nbr) |>
  arrange(encounter_id, .by_group = TRUE) |>
  mutate(visit_num = row_number()) |>
  ungroup() |>
  select(-encounter_id) |>
  pivot_wider(
    names_from = visit_num,                
    values_from = number_diagnoses,        
    names_prefix = "visit_",          
    values_fill = list(number_diagnoses = NA)) |>
  pivot_longer(cols = c(-patient_nbr),
               names_to = 'visit',
               values_to = 'diagnosis') |>
  drop_na(diagnosis) |>
  group_by(visit) |>
  summarise(mean_diagnosis = mean(diagnosis), no_people = n()) |>
  mutate(visit = str_extract(visit, "\\d+"),
         visit = factor(visit,levels=sort(as.numeric(visit)))) |>
  ggplot(aes(x=visit,y=mean_diagnosis)) +
  geom_point(aes(color=no_people),size=3) +
  scale_color_gradient2(low='hotpink',
                     mid='orange',
                     high='cornflowerblue',
                     midpoint=12000,
                     limits=c(2,NA),
                     n.breaks=5,
                     transform = 'sqrt',
                     na.value = 'forestgreen',
                     guide = guide_colorbar(title = "Amount of people")
                     ) +
  theme_minimal() +
  theme(axis.text.x=element_text(angle=90,hjust=1)) + 
  labs(title = 'Average number of diagnoses per visit by patient',
       subtitle = '- Green dots signify 1 person',
       x = 'Number of visit by patients',
       y = 'Average number of diagnoses') 
    
```

## Most common primary diagnoses compared to number of visits

```{r most_common_primary_diagnoses_per_visit}
data |>
  select(patient_nbr, encounter_id, nr_encounters, diag_1_long) |>
  group_by(diagnosis = diag_1_long) |>
  summarise(count = n(), .groups = "drop") |>
  top_n(10, count) |>
  inner_join(data, by = c("diagnosis" = "diag_1_long")) |>
  group_by(nr_encounters, diagnosis) |>
  summarise(diagnosis_count = n(), .groups = "drop") |>
  ggplot(aes(x = nr_encounters, y = diagnosis, fill = diagnosis_count)) +
  geom_tile(color = "white") +
  scale_fill_gradient(low = "orange", high = "purple") +
  scale_y_discrete(labels = function(y) str_wrap(y, width = 45)) +
  labs(
    x = "Number of visits",
    y = "",
    fill = "Amount of patients",
    title = "Top 10 most common primary diagnosis per visit"
  ) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 0, hjust = 1),
        legend.position = 'bottom',
        legend.background = element_rect(linewidth = 0.5, color = 'grey'),
        legend.text = element_text(angle=20,hjust=1),
        legend.justification = 'left',
        legend.title.position = 'top',
        plot.title = element_text(hjust = -3))
```

### Gender discrepancy for patients

```{r}
data |>
  group_by(patient_nbr) |>
  filter(n_distinct(gender) > 1) |>
  arrange(encounter_id, .by_group = TRUE)
```

## Diabetes types

```{r}
diabetes_count <- data |>
  select(patient_nbr,ends_with('_long')) |>
  pivot_longer(cols = ends_with('_long'),
               values_to = 'all_diagnoses',
               names_to = 'diag_order') |>
  select(-diag_order) |>
  group_by(patient_nbr) |>
  filter(str_detect(all_diagnoses,'iabetes')) |>
  distinct() |>
  group_by(all_diagnoses) |>
  summarise(count=n()) 
```

```{r diabetes_management_and_complications_for_all_patients}

diabetes_count |>
  mutate(type = ifelse(str_detect(all_diagnoses,'type I '),'Type I','Type II or unspecified'),
         uncontrolled = ifelse(str_detect(all_diagnoses,', uncontrolled'), 
                               'Uncontrolled', 'Not specified as uncontrolled'),
         no_complication = ifelse(str_detect(all_diagnoses, 'without mention of complication'),
                                  'No complication','Complications')) |>
  mutate(complication_type = ifelse(no_complication == 'Complications',
                                    str_to_sentence(str_extract(all_diagnoses,'(?<=with ).*(?=, *type)')),
                                    'None')) |>
  mutate(complication_type = ifelse(is.na(complication_type),'Pregnancy complications',complication_type)) |>
  mutate(complication_type = factor(complication_type, 
                                    levels = c(unique(complication_type[complication_type != 'None']), 'None'))) |>
  ggplot(aes(x = uncontrolled, y=count, fill = complication_type)) +
  geom_bar(stat='identity',) +
  facet_wrap(~type) +
  labs(title = 'Diabetes managment and complications',
       subtitle = str_c('For patients with diabetes as primary or secondary diagnosis (',
                        sum(diabetes_count$count),' patients)'),
       x = 'Controllation of disease', 
       y = 'Number of instances', 
       fill = 'Complications') +
  theme_minimal() +
  scale_x_discrete(labels = function(x) str_wrap(x,width = 17)) +
  scale_fill_viridis_d(option = 'viridis',direction=-1) 

```

### Only for people with complications

```{r diabetes_management_and_complications_for_patients_with_complications}
diabetes_count |>
  mutate(type = ifelse(str_detect(all_diagnoses,'type I '),'Type I','Type II or unspecified'),
         uncontrolled = ifelse(str_detect(all_diagnoses,', uncontrolled'), 
                               'Uncontrolled', 'Not specified as uncontrolled'),
         no_complication = ifelse(str_detect(all_diagnoses, 'without mention of complication'),
                                  'No complication','Complications')) |>
  mutate(complication_type = ifelse(no_complication == 'Complications',
                                    str_to_sentence(str_extract(all_diagnoses,'(?<=with ).*(?=, *type)')),
                                    'None')) |>
  mutate(complication_type = ifelse(is.na(complication_type),'Pregnancy complications',complication_type)) |>
  filter(!complication_type == 'None') |>
  ggplot(aes(x = uncontrolled, y=count, fill = complication_type)) +
  geom_bar(stat='identity',) +
  facet_wrap(~type) +
  labs(title = 'Diabetes managment and complications',
       subtitle = str_c(
         'For patients with diabetes as primary or secondary diagnosis\n(only for cases with complications: ',
                        diabetes_count |>
                          filter(!str_detect(all_diagnoses, 
                                             'without mention of complication')) |>
                          select(count) |>
                          sum(),
                        ' patients)'),
       x = 'Controllation of disease', 
       y = 'Number of instances', 
       fill = 'Complications') +
  theme_minimal() +
  scale_fill_viridis_d(option = 'plasma',direction=-1) + 
  scale_x_discrete(labels = function(x) str_wrap(x,width = 17))

```
